requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
langchain-google-genai==0.0.6
python-dotenv==1.0.0
pydantic==1.10.13
httpx==0.25.2
jinja2==3.1.2
python-multipart==0.0.6
openai>=1.6.1,<2.0.0